requirements.txt
discord.py
discord.py[voice]
dotenv
youtube_dl